services/log/pyproject.toml
[tool.poetry]
name = "log-service"
version = "0.1.0"
description = "SentinelMesh log ingestion service"
authors = ["Yigit Yumus"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
sqlalchemy = "^2.0.0"
pydantic-settings = "^2.0.0"
psycopg = { extras = ["binary"], version = "^3.1.0" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
